añado la membresía para el modal del perfil
