+++ frontCripta/src/components/profile/profileModal.scss
@@ -373,6 +373,9 @@
   }
 
   &__cta {
+    width: 100%;
+    border: none;
+    cursor: pointer;
     display: flex;
     align-items: center;
     justify-content: center;
